lock focus
(1) you need to lock the focus/exposure/white-balance. When these settings are not locked the camera frequently updates and the color changes dramatically in iOS 8 this happens more often and faster than iOS 7 - this is why we are seeing this issue more on iOS 8. see code example here.

(2) the algorithm uses the first frame to learn the background. Currently the background removal is working immediately when the app is open. I think we should change this behavior as following: User opens the app (the camera is open but the foreground extraction is not working yet) -> The user selects background/silhouette -> the user places a user/object inside the silhouette and presses a "ready" button -> the silhouette image disappears, the app locks the focus/exposure/white-balance and the algorithm starts working -> when the user is ready he presses the record button.

I've added a "green button". pressing on this button will lock the focus, and remove the siluete